+++ CloudShell.ControlPlane.Providers/Runtime/ProviderExecution.cs
@@ -616,6 +616,9 @@ public static class ProviderExecutionInstructionTypes
     public const string ContainerApplicationRoutingTearDown = "containerApplication.routing.tearDown";
     public const string ContainerApplicationServiceInstanceStart = "containerApplication.serviceInstance.start";
     public const string ContainerApplicationServiceInstanceStop = "containerApplication.serviceInstance.stop";
+    public const string AspNetCoreProjectStart = "aspNetCoreProject.start";
+    public const string AspNetCoreProjectStop = "aspNetCoreProject.stop";
+    public const string AspNetCoreProjectRestart = "aspNetCoreProject.restart";
     public const string ExecutableApplicationStart = "executableApplication.start";
     public const string ProcessStart = "process.start";
     public const string ProcessStop = "process.stop";